Fixed incorrect typing in `BaseSetVoltagePage` in __init__ method

+++ beskar/pages.py
@@ -295,7 +295,7 @@ class BaseSetVoltagePage(BasePage):
         about_header: str,
         desc: str,
         desc_width,
-        spin_box_range: Tuple[Union[int, float], Union[int, float]],
+        spin_box_range: Tuple[Union[int, int], Union[float, float]],
         suffix: str = ' Volts'
     ):
         self.set_with_settings = False
